// src/types/firstVoices.ts
export type firstVoicesType =
  | '次の番号は'
  | 'はいはいー次はーー'
  | 'さてさて次の番号はーー'
  | 'さてさてお次はー'
  | '次はどれにしようかなんー'
  | '次は何かなえーと'
  | '次に選ぶのは'
  | '選ばれたのは'
  | '次はこの番号で決まり'
  | 'どれにしようかなえーと'
  | 'そうだな次の番号は'
  | '次はこれだね'
  | '次はどれがいいかな';